mas_GRDH/requirements.txt
torch>=1.13.0
torchvision>=0.14.0
pytorch-lightning>=1.9.0
omegaconf>=2.3.0
transformers>=4.25.0
diffusers>=0.14.0
einops>=0.6.0
numpy>=1.23.0
Pillow>=9.3.0
opencv-python>=4.7.0
scipy>=1.10.0
tqdm>=4.64.0
kornia>=0.6.0
